+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+- Handles git index version 4
+- Handles git index version 3 with extended flags properly
+
 # 0.1.0
 
 Initial release